Updated SignUp Page Connected Pages to Self-hosted MySQL server
Incorporated changes that the teacher wanted since last presentation.

Added integration with the self-hosted database that I'm running (my domain name is "Aetacraft.tk").

Fixed an issue where $_SESSION information was attempting to be read, failing, and then throwing an error up onto the webpage (which looked unprofessional).

I also cleaned up the code for the "signUp.php" page. As that was bothering me and it also fixed some of the issues I was having with the database.